Details page with Promise instead of Observables
This only calls for latest 20 Flickr photos on the Homepage, and the Details page will display the appropriate details of those images already pulled from Flickr, not the newest real-time photos.  I use the thumbnail index on the Homepage to get the corresponding details of the image the user selects.